separate htmlparser
